% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{center_data}
\alias{center_data}
\title{Generate centered versions of all variables in a data set.}
\usage{
center_data(data, cluster)
}
\arguments{
\item{data}{Data frame, matrix, or tibble containing the data to center.}

\item{cluster}{Column label of \code{data} corresponding to the cluster/group identification variable.}
}
\value{
An augmented database with variables indicating level-1 cluster means, cluster-mean centered level-1 variables, grand-mean centered level-1 variables, and grand-mean centered level-2 variables.
}
\description{
Variables that are cluster-mean centered will be labeled with a "cwc" suffix (abbreviation for centered within cluster),
the cluster means of level-1 variables will be labeled with a "means" suffix, and
variables that are grand-mean centered will be labeled with a "cgm" suffix (abbreviation for centered grand mean).
}
\examples{
\dontrun{
center_data(data = timss, cluster = "idteach")
}
}
