% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{center_within}
\alias{center_within}
\title{Center a vector of numeric values using cluster means.}
\usage{
center_within(x, cluster)
}
\arguments{
\item{x}{Numeric vector.}

\item{cluster}{Vector of cluster identifiers.}
}
\value{
A cluster-mean centered vector.
}
\description{
Center a vector of numeric values using cluster means.
}
