% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{chisq_hlmer}
\alias{chisq_hlmer}
\title{Chi-square estimates for random-effects variances}
\usage{
chisq_hlmer(model, summary = NULL, y_lvl1, cluster, x_lvl1 = NULL,
  x_lvl2 = NULL, data)
}
\arguments{
\item{model}{Raw lmer output model.}

\item{summary}{Optional lmer summary model.}

\item{y_lvl1}{Column label of \code{data} corresponding to the criterion variable for level-1 observations.}

\item{cluster}{Column label of \code{data} corresponding to the cluster/group identification variable.}

\item{x_lvl1}{Column label(s) of \code{data} corresponding to the predictor variable(s) for level-1 observations.}

\item{x_lvl2}{Column label(s) of \code{data} corresponding to the predictor variable(s) for level-2 observations.}

\item{data}{Data frame, matrix, or tibble containing the data to use in the linear model.}
}
\value{
Table of chi-square estimates for random-effects variances.
}
\description{
Chi-square estimates for random-effects variances
}
\author{
Jeffrey Dahlke and Jonathan Brown
}
