% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cluster_means}
\alias{cluster_means}
\title{Obtain cluster means from a numeric vector.}
\usage{
cluster_means(x, cluster)
}
\arguments{
\item{x}{Numeric vector.}

\item{cluster}{Vector of cluster identifiers.}
}
\value{
A vector of cluster means.
}
\description{
Obtain cluster means from a numeric vector.
}
