% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hsb}
\alias{hsb}
\title{Level-1 and level-2 data from the High School and Beyond Survey}
\format{data.frame}
\usage{
data(hsb)
}
\description{
Level-1 and level-2 data from the High School and Beyond Survey
}
\examples{
data(hsb)
}
\keyword{datasets}
