% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hsb1}
\alias{hsb1}
\title{Level-1 data from the High School and Beyond Survey}
\format{data.frame}
\usage{
data(hsb1)
}
\description{
Level-1 data from the High School and Beyond Survey
}
\examples{
data(hsb1)
}
\keyword{datasets}
