% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hsb2}
\alias{hsb2}
\title{Level-2 data from the High School and Beyond Survey}
\format{data.frame}
\usage{
data(hsb2)
}
\description{
Level-2 data from the High School and Beyond Survey
}
\examples{
data(hsb2)
}
\keyword{datasets}
