% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{print.hlmer.hlmerMod}
\alias{print.hlmer.hlmerMod}
\title{Print method for hlmerMod-class objects}
\usage{
\method{print}{hlmer.hlmerMod}(x, ..., digits = 5)
}
\arguments{
\item{x}{hlmerMod-class object.}

\item{...}{Additional arguments for \code{print()}.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
A formatted hlmerMod-class object.
}
\description{
Print method for hlmerMod-class objects
}
