% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{print.hlmer}
\alias{print.hlmer}
\title{Print method for hlmer-class objects}
\usage{
\method{print}{hlmer}(x, ..., digits = 5)
}
\arguments{
\item{x}{hlmer-class object.}

\item{...}{Additional arguments for \code{print()}.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
A formatted hlmer-class object.
}
\description{
Print method for hlmer-class objects
}
