% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{rel_lvl1}
\alias{rel_lvl1}
\title{Estimate the reliability of random effects}
\usage{
rel_lvl1(summary, cluster, x_lvl1 = NULL, data)
}
\arguments{
\item{summary}{lmer summary model.}

\item{cluster}{Column label of \code{data} corresponding to the cluster/group identification variable.}

\item{x_lvl1}{Column label(s) of \code{data} corresponding to the predictor variable(s) for level-1 observations.}

\item{data}{Data frame, matrix, or tibble containing the data to use in the linear model.}
}
\value{
Vector of reliability estimates for random effects.
}
\description{
Estimate the reliability of random effects
}
