% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmer.R
\name{se_lvl1}
\alias{se_lvl1}
\title{Estimate level-1 standard errors for random effects}
\usage{
se_lvl1(sigma, cluster, x_lvl1 = NULL, data)
}
\arguments{
\item{sigma}{Residual standard deviation.}

\item{cluster}{Column label of \code{data} corresponding to the cluster/group identification variable.}

\item{x_lvl1}{Column label(s) of \code{data} corresponding to the predictor variable(s) for level-1 observations.}

\item{data}{Data frame, matrix, or tibble containing the data to use in the linear model.}
}
\value{
Matrix of estimated level-1 standard errors for random effects.
}
\description{
Estimate level-1 standard errors for random effects
}
